requirements.txt
boto==2.39.0
dj-database-url==0.3.0
Django==1.9.1
django-allauth==0.24.1
django-appconf==1.0.1
django-crispy-forms==1.6.0
django-imagekit==3.3
django-storages-redux==1.3.1
gunicorn==19.4.5
oauthlib==1.0.3
pilkit==1.1.13
Pillow==3.1.0
psycopg2==2.6.1
python-openid==2.2.5
requests==2.9.1
requests-oauthlib==0.6.0
rq==0.5.6
six==1.10.0
wheel==0.24.0
